Correct errors related to supporting v1.34.26 and later
Code contributed to support de1app v1.34.26 and later
caused two types of errors:

* domain error: argument not in valid range

* Errors when running using the simulator without a DE1

The first of these appears to have come from a hard-to-replicate
condition where both the flow and the pressure during preinfusion or
pouring are both zero.  An explicit check is now made for $GroupFlow
or $GroupPressure (or their $::de1() counterparts) being zero or negative,
returning 0 for the "puck resistance" in these cases.

The off-line errors revealed that SetMixTemp had been used
instead of MixTemp. This has been corrected.

The check for "running with the new code base" has been soimplified to

  [ package vsatisfies [package version de1app] 1.34.26- ]

DSx version bump to 4.47

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>